% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesdir-package.R
\docType{package}
\name{bayesdir-package}
\alias{bayesdir}
\alias{bayesdir-package}
\title{bayesdir}
\description{
This package implements the \code{bayesdir} method.
}
\author{
\strong{Maintainer}: Sheng Jiang \email{jiangsheng@cuhk.edu.cn}

}
\keyword{internal}
