% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{compute_weight_sum}
\alias{compute_weight_sum}
\title{Compute the sum of weights sum_{t = start}^{end} 1/(t + N)^2.}
\usage{
compute_weight_sum(start, end, N)
}
\arguments{
\item{start}{starting index (t >= 1)}

\item{end}{ending index}

\item{N}{number of observed data points}
}
\value{
scalar value of the sum
}
\description{
Compute the sum of weights sum_{t = start}^{end} 1/(t + N)^2.
}
