% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{compute_weights}
\alias{compute_weights}
\title{Compute the sum of weights sum_{t = start}^{end} 1/(t + N)^2.}
\usage{
compute_weights(n, M)
}
\arguments{
\item{n}{number of observed data points}

\item{M}{number of iterations}
}
\description{
Compute the sum of weights sum_{t = start}^{end} 1/(t + N)^2.
}
