% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{generate_vmf_data}
\alias{generate_vmf_data}
\title{generate n vmf data when p=2}
\usage{
generate_vmf_data(n, mu_true, kappa_true)
}
\arguments{
\item{n}{number of observations}

\item{mu_true, kappa_true}{true parameter}
}
\value{
list containing both cartesian and polara coodinates type.
}
\description{
generate n vmf data when p=2
}
