% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{get_Ap}
\alias{get_Ap}
\title{Compute the function A(kappa) = I_{p/2}(kappa) / I_{p/2-1}(kappa)
for the von Mises–Fisher distribution.}
\usage{
get_Ap(kappa, p = 2)
}
\arguments{
\item{kappa}{concentration parameter (non‑negative scalar)}

\item{p}{space dimension (default = 2)}
}
\value{
scalar value of A(kappa)
}
\description{
Compute the function A(kappa) = I_{p/2}(kappa) / I_{p/2-1}(kappa)
for the von Mises–Fisher distribution.
}
