% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{get_FIM_vmf}
\alias{get_FIM_vmf}
\title{Fisher information matrix for the vMF model}
\usage{
get_FIM_vmf(C, p = length(C))
}
\arguments{
\item{C}{natural parameter vector (length p)}

\item{p}{dimension (defaults to length(C))}
}
\value{
p x p Fisher information matrix
}
\description{
Fisher information matrix for the vMF model
}
