% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{get_stats}
\alias{get_stats}
\title{Helper to compute statistics (Coverage, Length, Variance)}
\usage{
get_stats(samples, true_val)
}
\arguments{
\item{samples}{samples of parameters on the real line}

\item{true_val}{ground truth on real line}
}
\value{
list contain stats e,g, cover or not, len of confidence interval, variance.
}
\description{
Helper to compute statistics (Coverage, Length, Variance)
}
