% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{mle_vmf}
\alias{mle_vmf}
\title{Compute MLE given data X}
\usage{
mle_vmf(X)
}
\arguments{
\item{X}{nrow = p ncol = number of observation}
}
\value{
MLE as a p-dim vector
}
\description{
Compute MLE given data X
}
