% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.R
\name{run_early_path}
\alias{run_early_path}
\title{Run the early path for a single chain (Step 1 of Hybrid Method)}
\usage{
run_early_path(c0, n, M)
}
\arguments{
\item{c0}{initial natural parameter (MLE)}

\item{n}{observed sample size}

\item{M}{predictive steps}
}
\value{
A list containing displacement D and quadratic variation Q
}
\description{
Run the early path for a single chain (Step 1 of Hybrid Method)
}
