% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_early_path.R
\name{run_hybrid_vmf}
\alias{run_hybrid_vmf}
\title{Run Hybrid Predictive Resampling for vMF}
\usage{
run_hybrid_vmf(X, B = 5000, M = 50, n_cores = 1)
}
\arguments{
\item{X}{data matrix (p x n)}

\item{B}{number of independent chains}

\item{M}{number of early predictive steps}

\item{n_cores}{number of cores for parallel processing}
}
\description{
Run Hybrid Predictive Resampling for vMF
}
