% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.R
\name{run_predresamp_vmf}
\alias{run_predresamp_vmf}
\title{run B chains for empirical density}
\usage{
run_predresamp_vmf(X, B = 5000, M = 50, n_cores = 1, return_chains = FALSE)
}
\arguments{
\item{X}{data dim = c(p,n)}

\item{B}{expected sample size}

\item{M}{number of iteration}

\item{n_cores}{number of cores used in mclapply}

\item{save_chain}{whether to save the entire chain.}
}
\description{
run B chains for empirical density
}
