% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{rvMF2}
\alias{rvMF2}
\title{Draw a random unit vector from the vMF(eta) distribution on the circle. Used in simulation for comparison,
Only implemented for p = 2.}
\usage{
rvMF2(eta)
}
\arguments{
\item{eta}{natural parameter vector (length 2)}
}
\value{
2‑dimensional unit vector
}
\description{
Draw a random unit vector from the vMF(eta) distribution on the circle. Used in simulation for comparison,
Only implemented for p = 2.
}
