% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{safe_sym_inverse}
\alias{safe_sym_inverse}
\title{compute inverse for squared matrix}
\usage{
safe_sym_inverse(M, tol = 1e-10)
}
\arguments{
\item{M}{matrix}

\item{tol}{tolerence erro}
}
\value{
M^{-1}
}
\description{
compute inverse for squared matrix
}
