% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{score_fn}
\alias{score_fn}
\title{Compute the score function s(eta, x) = x - E\link{X} for vMF.}
\usage{
score_fn(eta, x, p = 2)
}
\arguments{
\item{eta}{natural parameter vector (length p)}

\item{x}{unit vector observation (length p)}

\item{p}{space dimension (default = 2)}
}
\value{
score vector of length p
}
\description{
Compute the score function s(eta, x) = x - E\link{X} for vMF.
}
