% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{sim_martingale_study}
\alias{sim_martingale_study}
\title{functions for simulations study}
\usage{
sim_martingale_study(
  n,
  n_sim = 500,
  mu_true = 2,
  kappa_true = 4,
  B = 5000,
  M = 50,
  seed = 1234,
  n_cores = 1,
  save_all = FALSE
)
}
\arguments{
\item{n}{sample size}

\item{n_sim}{sim times}

\item{B}{number of chains}

\item{M}{number of iteration for each chain}

\item{seed}{for reproducability}

\item{n_cores}{number of cores you can use}

\item{save_all}{save data or not}

\item{mu_ture, kappa_true}{true param}
}
\value{
all the data with a summary we gonna used
}
\description{
functions for simulations study
}
