% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.R
\name{simulate_raw_path}
\alias{simulate_raw_path}
\title{Simulate Raw Score Path (Step 1 of Algorithm 1)}
\usage{
simulate_raw_path(c0, n, M, save_raw = FALSE)
}
\arguments{
\item{c0}{Initial estimator (MLE)}

\item{n}{Initial sample size}

\item{M}{Depth (iterations)}

\item{save_raw}{Whether to save the trajectory of c_n,m}
}
\value{
A list containing the raw difference D, quadratic variation Q, and optional path
}
\description{
Simulate Raw Score Path (Step 1 of Algorithm 1)
}
