% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{summarize_df}
\alias{summarize_df}
\title{Final Summarization}
\usage{
summarize_df(res_list, method_name)
}
\arguments{
\item{method_name}{for example, "MCMC".}

\item{rel_list}{a list containing all samples from all simulations}
}
\value{
list containing all statistics
}
\description{
Final Summarization
}
