% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{compute_weights}
\alias{compute_weights}
\title{Helper to compute weights sums}
\usage{
compute_weights(n, M)
}
\arguments{
\item{n}{number of observed data points}

\item{M}{number of iterations}
}
\value{
weights of accumlated and remaining
}
\description{
Helper to compute weights sums
}
