% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp.R
\name{run_predresamp_vmf}
\alias{run_predresamp_vmf}
\title{Score-Only Martingale Posterior Sampler for vMF}
\usage{
run_predresamp_vmf(
  X,
  B = 5000,
  M = 50,
  type = c("HRS-PC", "RS-PC"),
  cal = c("pooled", "pathwise"),
  n_cores = 1,
  save_raw = FALSE
)
}
\arguments{
\item{X}{Data matrix (p x n)}

\item{B}{Number of chains}

\item{M}{Depth M}

\item{type}{"RS-PC" or "HRS-PC"}

\item{cal}{"pooled" or "pathwise"}

\item{n_cores}{Number of cores}

\item{save_raw}{Whether to return the raw score paths}
}
\description{
Score-Only Martingale Posterior Sampler for vMF
}
